+++ CHANGES.md
@@ -1,5 +1,6 @@
 ### v6.0.2 (2021-10-27)
 
+* dns: remove astring dependency (@hannesm)
 * dns-client: adapt to happy-eyeballs 0.1.0 API (#274 @hannesm)
 * dns-client: avoid List.concat_map, make it available on OCaml 4.08 (@hannesm)